+++ testng.xml
@@ -12,7 +12,7 @@
             <class name="com.softserve.edu.greencity.api.tests.econews.negative.CreateNewsWithDetailedResponseTest"/>
             <class name="com.softserve.edu.greencity.api.tests.econews.negative.CreateNewsWithShortResponseTest"/>
 
-            <class name="com.softserve.edu.greencity.api.tests.signin.SignInApiTest"/>
+            <class name="com.softserve.edu.greencity.api.tests.ownsecurity.signin.SignInApiTest"/>
         </classes>
     </test>
